remove size controller
